Disable form on submit

+++ static/js/main.js
@@ -1,43 +1,46 @@
 let form = document.getElementById('form')
 let status = document.getElementById('status')
 let fileList = document.getElementById('file-list')
-let btn = document.getElementById('btn')
+let disabled = false
 
 form.onsubmit = e => {
-	btn.disabled = true
-	let data = new FormData(form)
-	axios.post('/upload', data, {
-		onUploadProgress: e => {
-			let progress = parseInt(((e.loaded / e.total) * 100))
-			status.innerText = `Uploading ${progress}%`
-		}
-	})
-	.then(response => {
-		btn.disabled = false
-		status.innerText = `Uploaded`
-		reload()
-	})
-	.catch(err => {
-		btn.disabled = false
-		status.innerText = `Error: ${err.response.data}`
-	})
+	if(!disabled){
+		disabled = true
+		let data = new FormData(form)
+		axios.post('/upload', data, {
+			onUploadProgress: e => {
+				let progress = parseInt(((e.loaded / e.total) * 100))
+				status.innerText = `Uploading ${progress}%`
+			}
+		})
+		.then(response => {
+			disabled = false
+			status.innerText = `Uploaded`
+			reload()
+		})
+		.catch(err => {
+			disabled = false
+			status.innerText = `Error: ${err.response.data}`
+		})
+	}
 	e.preventDefault()
 }
 
-function remove(e){
-	let file = e.target.dataset.file
-	e.target.disabled = true
-	status.innerText = `Deleting ...`
-	axios.delete(`/delete/${file}`)
-	.then(() => {
-		e.target.disabled = false
-		status.innerText = `Deleted`
-		reload()
-	})
-	.catch(err => {
-		e.target.disabled = false
-		status.innerText = `Error: ${err.response.data}`
-	})
+function remove(file){
+	if(!disabled){
+		disabled = true
+		status.innerText = `Deleting ...`
+		axios.delete(`/delete/${file}`)
+		.then(() => {
+			disabled = false
+			status.innerText = `Deleted`
+			reload()
+		})
+		.catch(err => {
+			disabled = false
+			status.innerText = `Error: ${err.response.data}`
+		})
+	}
 }
 
 function reload(){
@@ -45,7 +48,7 @@ function reload(){
 	.then(response => {
 		fileList.innerHTML = ''
 		response.data.forEach(item => {
-			fileList.innerHTML += `<p><button data-file="${item}" onclick="remove(event)">Delete</button>&emsp;<a href="/download/${item}">${item}</a></p>`
+			fileList.innerHTML += `<p><button onclick="remove('${item}')">Delete</button>&emsp;<a href="/download/${item}">${item}</a></p>`
 		})
 	})
 	.catch(err => {
